Redid some mob spawning methods, rebirth is a lot simpler (doesnt use a runnable). Added SpawnMethod enum

+++ src/main/java/deimophobe/nightfall/Game.java
@@ -110,8 +110,8 @@ public class Game {
 	private Plague activePlague = null;
 
 
-	private long tickNumber = 0;
-	public long getCurrentTick() { return tickNumber; }
+	private int tickNumber = 0;
+	public int getCurrentTick() { return tickNumber; }
 
 	private Game(GameMap map) {
 		game = this;
